third_party/WebKit/LayoutTests/fast/dom/Window/getMatchedCSSRules-parent-stylesheets-expected.txt

CONSOLE WARNING: line 16: document.getMatchedCSSRules() is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4606972603138048 for more details.
getMatchedCSSRules returns CSSStyleRules with correct parent stylesheet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


parentStyleSheet media: ''
parentStyleSheet media: '(min-device-width: 120px)'
parentStyleSheet media: '(min-device-width: 20px)'
parentStyleSheet media: '(min-device-width: 220px)'
PASS successfullyParsed is true

TEST COMPLETE
Test getMatchedCSSRules to correctly setup CSSStyleRule parent stylesheet.

This span should have border and be written in red.
